support Graph in raw format

+++ test/source/test_cycle_ratio.cpp
@@ -14,14 +14,14 @@
 #include <xnetwork/generators/testcases.hpp> // for create_test_case1,
 
 TEST_CASE("Test Cycle Ratio") {
-  const auto indices = std::array<uint32_t, 5>{0, 1, 2, 3, 4};
+  const auto indices = std::array<size_t, 5>{0, 1, 2, 3, 4};
   auto gra = create_test_case1(indices);
 
   const auto cost = std::array<int, 5>{5, 1, 1, 1, 1};
 
   const auto get_cost = [&](const auto &edge) -> int {
     const auto [u, v] = edge;
-    return cost[gra[u][v]];
+    return cost[size_t(gra[u][v])];
   };
   const auto get_time = [&](const auto & /*e*/) -> int { return 1; };
 
@@ -37,14 +37,14 @@ TEST_CASE("Test Cycle Ratio") {
 TEST_CASE("Test Cycle Ratio of Timing Graph") {
   // make sure no parallel edges!!!
 
-  const auto indices = std::array<uint32_t, 6>{0, 1, 2, 3, 4, 5};
+  const auto indices = std::array<size_t, 6>{0, 1, 2, 3, 4, 5};
   auto gra = create_test_case2(indices);
 
   const auto cost = std::array<int, 6>{7, -1, 3, 0, 2, 4};
 
   const auto get_cost = [&](const auto &edge) -> int {
     const auto [u, v] = edge;
-    return cost[gra[u][v]];
+    return cost[size_t(gra[u][v])];
   };
   const auto get_time = [&](const auto & /*e*/) -> int { return 1; };
